FixBugAndSomeChange

+++ Shop.Aplication/Commands/CartCommand/DeleteCartCommand.cs
@@ -1,4 +1,6 @@
 ﻿using MediatR;
+using Microsoft.AspNetCore.Http;
+using Shop.Domain.Ultils;
 using Shop.Infratructure.UnitOfWork;
 
 namespace Shop.Aplication.Commands.CartCommand;
@@ -16,19 +18,22 @@ public class DeleteCartCommand:IRequest<bool>
 public class HandDeleteCartCommand:IRequestHandler<DeleteCartCommand,bool>
 {
     private readonly IUnitOfWork _unitOfWork;
-
-    public HandDeleteCartCommand(IUnitOfWork unitOfWork)
+    private readonly IHttpContextAccessor _contextAccessor;
+    public HandDeleteCartCommand(IUnitOfWork unitOfWork, IHttpContextAccessor contextAccessor)
     {
         _unitOfWork = unitOfWork;
+        _contextAccessor = contextAccessor;
     }
 
     public async Task<bool> Handle(DeleteCartCommand request, CancellationToken cancellationToken)
     {
         try
         {
+            var UserId = Guid.Parse(_contextAccessor.HttpContext!.User.GetIdFromClaim());
+
             var cart = await _unitOfWork.cartRepository.FindByIdAsync(request.Id);
 
-            if (cart is null) return false;
+            if (cart is null || cart.UserId.Equals(UserId) is false) return false;
         
             _unitOfWork.cartRepository.Remove(cart);
         
@@ -36,9 +41,9 @@ public class HandDeleteCartCommand:IRequestHandler<DeleteCartCommand,bool>
         
             return true;
         }
-        catch (Exception e)
+        catch (Exception )
         {
-            throw new Exception(e.Message);
+            throw ;
         }
        
     }
