+++ Shop.Aplication/Commands/CommentCommand/DeleteCommentCommand.cs
@@ -1,4 +1,6 @@
 ﻿using MediatR;
+using Microsoft.AspNetCore.Http;
+using Shop.Domain.Ultils;
 using Shop.Infratructure.UnitOfWork;
 
 namespace Shop.Aplication.Commands.CommentCommand;
@@ -15,20 +17,22 @@ public class DeleteCommentCommand:IRequest<bool>
 public class HandCommentCommand:IRequestHandler<DeleteCommentCommand,bool>
 {
     private readonly IUnitOfWork _unitOfWork;
-
-    public HandCommentCommand(IUnitOfWork unitOfWork)
+    private readonly IHttpContextAccessor _contextAccessor;
+    public HandCommentCommand(IUnitOfWork unitOfWork, IHttpContextAccessor contextAccessor)
     {
         _unitOfWork = unitOfWork;
+        _contextAccessor = contextAccessor;
     }
 
     public async Task<bool> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
     {
         try
         {
-            
+            var UserId = Guid.Parse(_contextAccessor.HttpContext!.User.GetIdFromClaim());
+
             var comment = await _unitOfWork.commentRepository.FindByIdAsync(request.Id);
 
-            if (comment is null ) return false;
+            if (comment is null || comment.UserId != UserId ) return false;
         
             _unitOfWork.commentRepository.Remove(comment); 
         
@@ -36,9 +40,9 @@ public class HandCommentCommand:IRequestHandler<DeleteCommentCommand,bool>
 
             return true;
         }
-        catch (Exception e)
+        catch (Exception )
         {
-            throw new Exception(e.Message);
+            throw ;
         }
 
     }
